rewrite ChangingProgressbar to ChangingProgressbarProvider
